Created base schedule page
All functions as described in the prototype have been implemented, i.e.: Expanding to see speakers, changing the time zones.
Aesthetics, on the other hand, could be improved. The most notable issue right now is making it obvious to users that the table can be expanded.

Any other areas of improvement to recommend?